Increase blockwise timeouts; add waitACK to proxy config; handle empty response bodies
Large sync responses take longer than the default `time.Second` to send over blockwise
xfer so increase the timeout.

Added `WaitTimeBeforeACK` to the proxy `Config` which is
the number of seconds to wait before sending an ACK back if the upstream HS hasn't
responded yet. Without this, large sync responses which take a while to calculate on
the HS side would cause the client to retransmit the sync request, which, when coupled
with a go-coap bug https://github.com/plgd-dev/go-coap/issues/218 would cause the
large response to be sent multiple times via blockwise xfer.

Sometimes servers respond with no bodies, which was not correctly handled in the proxy
server or some CLI tools.